refactor(pre-tool-use): update import paths for helper modules
- Refactored import paths for various hooks to reflect their new locations within the pre-tool-use helpers directory.
- Ensured consistency and clarity in the import statements across multiple test files, enhancing code organization.
