+++ packages/filid/src/__tests__/unit/hooks/pre-tool-validator.test.ts
@@ -1,9 +1,7 @@
 import { describe, expect, it } from 'vitest';
 
-import {
-  isDetailMd,
-  validatePreToolUse,
-} from '../../../hooks/pre-tool-validator/pre-tool-validator.js';
+import { validatePreToolUse } from '../../../hooks/pre-tool-use/helpers/pre-tool-validator/pre-tool-validator.js';
+import { isDetailMd } from '../../../hooks/shared/shared.js';
 import type { PreToolUseInput } from '../../../types/hooks.js';
 
 const baseInput: PreToolUseInput = {
